[GEOS-11525] GeoFence: refactor cache classes
